Initial BigQuery query runner implemented. First study tests passed.
