+++ dae/dae/inmemory_storage/raw_variants.py
@@ -33,6 +33,7 @@ class RawVariantsQueryRunner(QueryRunner):
         assert self.variants_iterator is not None
 
     def run(self):
+        assert self._result_queue is not None
         try:
             if self.closed():
                 return
@@ -95,7 +96,7 @@ class RawFamilyVariants(abc.ABC):
             for fv in fvs:
                 for fa in fv.alt_alleles:
                     seen_in_status[fa.allele_index] = reduce(
-                        lambda t, s: t | s.value,
+                        lambda t, s: t | s.value,  # type: ignore
                         filter(None, fa.allele_in_statuses),
                         seen_in_status[fa.allele_index])
                     seen_as_denovo[fa.allele_index] = reduce(
@@ -260,7 +261,7 @@ class RawFamilyVariants(abc.ABC):
             genes=None,
             effect_types=None,
             variant_type=None,
-            person_ids=None,
+            # person_ids=None,
             **_kwargs):
         # pylint: disable=too-many-return-statements,too-many-branches
         """Return True if a summary allele meets the required conditions."""
@@ -282,11 +283,11 @@ class RawFamilyVariants(abc.ABC):
         if variant_type is not None:
             if not variant_type.match([allele.allele_type]):
                 return False
-        if person_ids is not None:
-            if allele.is_reference_allele:
-                return False
-            if not set(allele.variant_in_members) & set(person_ids):
-                return False
+        # if person_ids is not None:
+        #     if allele.is_reference_allele:
+        #         return False
+        #     if not set(allele.variant_in_members) & set(person_ids):
+        #         return False
         return True
 
     @classmethod
