+++ gcp_genotype_storage/gcp_genotype_storage/tests/test_big_query_study_queries.py
@@ -1,0 +1,37 @@
+# pylint: disable=W0621,C0114,C0116,W0212,W0613
+import pytest
+
+from dae.utils.regions import Region
+
+
+@pytest.fixture(scope="session")
+def bq_backend(imported_study):
+    return imported_study._backend
+
+
+@pytest.mark.parametrize("index,query,ecount", [
+    (1, {}, 2),
+    (2, {"genes": ["g1"]}, 2),
+    (3, {"genes": ["g2"]}, 0),
+    (4, {"effect_types": ["missense"]}, 1),
+    (5, {"effect_types": ["splice-site"]}, 1),
+    (6, {"regions": [Region("foo", 10, 10)]}, 1),
+])
+def test_family_queries(imported_study, index, query, ecount):
+    vs = list(imported_study.query_variants(**query))
+    assert len(vs) == ecount
+
+
+@pytest.mark.parametrize("index,query,ecount", [
+    (1, {}, 2),
+    (2, {"genes": ["g1"]}, 2),
+    (3, {"genes": ["g2"]}, 0),
+    (4, {"effect_types": ["missense"]}, 1),
+    (5, {"effect_types": ["splice-site"]}, 1),
+    (6, {"regions": [Region("foo", 10, 10)]}, 1),
+])
+def test_summary_queries(imported_study, index, query, ecount):
+
+    vs = list(imported_study.query_summary_variants(**query))
+
+    assert len(vs) == ecount
